+++ src/whatsapp/dto/instance.dto.ts
@@ -2,4 +2,5 @@ export class InstanceDto {
   instanceName: string;
   webhook?: string;
   events?: string[];
+  qrcode?: boolean;
 }